feat: store message receipt confirmations and timing in database
- Add message_receipts table to track per-peer receipt confirmations
- Add save_receipt() and load_receipts() functions in messages.rs
- Load receipts on TUI startup and populate broadcast_receipts and dm_receipts
- Save receipts when SwarmEvent::Receipt is received
- Update migrations to include all required columns (msg_id, sent_at, etc.)
- Fix diesel_generate.sh to output to src/generated/ and fix sed pattern
- Fix build.rs to handle Double and Nullable types correctly
- Add missing msg_id field in test

+++ migrations/2026-04-28-000000_message_receipts/up.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS message_receipts (
     msg_id TEXT NOT NULL,
     peer_id TEXT NOT NULL,
     kind INTEGER NOT NULL, -- 0=broadcast, 1=dm
-    confirmed_at REAL NOT NULL,
+    confirmed_at DOUBLE NOT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     UNIQUE (msg_id, peer_id, kind)
 );
